pyproject.toml
[project]
name = "media-cleaning-mvp"
version = "0.1.0"
description = "Audio/video transcription and cleaning pipeline MVP"
authors = [{ name = "Your Name" }]
requires-python = ">=3.10"
dependencies = [
  "faster-whisper>=1.0.0",
  "ffmpeg-python>=0.2.0",
  "soundfile>=0.12.1",
  "pydub>=0.25.1",
  "numpy>=1.26",
  "pandas>=2.2",
  "langdetect>=1.0.9",
  "unidecode>=1.3.8",
  "typer[all]>=0.9.0",
  "rich>=13.7",
  "sqlmodel>=0.0.16",
  "pyyaml>=6.0.1"
]

[project.optional-dependencies]
develop = [
  "pytest>=7.4",
  "mypy>=1.8"
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"
